requirements.txt
requests>=0.13.1
python-memcached>=1.48
